Slightly smarter game implemented.
The computer will now no longer select at random.
First, it checks if it can win on this move. If yes, then it makes the move.

Then it checks if the opponent can win in one move. If yes, then it blocks the opponent's move.

If neither hold true, then it makes a random move.